+++ CHANGELOG.rst
@@ -21,6 +21,7 @@ Added
     - App alerts for zone owner for zone actions (#1204)
     - ``member_notify_move`` app setting (#1203)
     - Project member notifications from zone moving (#1203)
+    - ``LandingZone.user_message`` field (#1203)
 - **Samplesheets**
     - Simple link support for string cell rendering (#1001)
     - ``generic_raw`` assay plugin (#1128)